Move all game logic-related files into a single folder
Someday I might want to be able to run games from the terminal. Today is not that day but it should now be a lot easier in the future.

Still need to figure out what to do with card parsing and such. I think I misunderstood how composables are intended to be used.