Rework attack structure from the ground up
Goals: to expose more metadata of the attack to the game core, and to allow more flexibility in parsing.

For example, for an attack with a conditional or modifier, the PTCGP app will calculate and display the correct base damage based on the current state of the game. This was impossible in the previous attack system, as attack effects were a complete black box. This update fixes that by adding a method that allows attacks to expose their calculation for base damage.

The straw that broke the camel's back was A2 Mesprit, whose attack can only be used when Uxie and Azelf are both on the bench. In the previous attack system, the best I could have done would have been to allow the attack to start but immediately quit with "attack failed". After this update, I can instead have the parser add a check through the extraConditions property, and the game will prevent the attack from being used if the check fails.

Finally, the last major change is breaking out coin flips from the black box, similar to base damage calculation. This allows the parser to flow through the attack text one phrase at a time, rather than having to look at the full text and recursively parse smaller parts of it to understand the whole structure.