Add attack pre-damage effect: remove tools from defending mon
- A2 Starly

Didn't realize I'd need to add a whole new type of effect for this one. Couldn't be a regular side effect because then Rocky Helmet would still hit back for 20, for example.

I tested and found out that fortunately, the app does not check for knockouts after removing the tool (the damage will still happen even if the defender's HP has already been reduced to 0).