Add ability triggers related to attaching/removing energy
- No retreat cost if energy attached: A2 Giratina

Rearranged a bunch of methods and added comments in the Game class because I'm tired of feeling lost in there. It's over 1000 lines now and I don't want to go much higher if I can avoid it, but at least now it's easier to navigate.

Also fixed a few small bugs of varying age, such as Item cards not being able to be played after a Supporter and damage-reduction player statuses getting applied to *both* players' mon.